4 - format code and fix init command
