+++ src/models/app/app.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""App models."""
 from dataclasses import dataclass
 from typing import List
 
@@ -7,6 +8,8 @@ from models.app.widget_kind import WidgetKind
 
 @dataclass
 class AppModel:
+    """Model for app controller."""
+
     title: str
     active_tab_index: int
     tab_widgets: List[WidgetKind]
