+++ src/models/weather/daily_forecast_model.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Weather models."""
 import datetime
 from dataclasses import dataclass
 
@@ -7,5 +8,7 @@ from models.weather.weather_forecast_data import WeatherForecastData
 
 @dataclass
 class DailyForecastModel:
+    """Model for daily weather forecast."""
+
     ts: datetime.date
     forecast: WeatherForecastData
