+++ src/models/weather/weather_forecast.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-
+"""Weather models."""
 from dataclasses import dataclass
 from typing import List
 
@@ -9,5 +9,7 @@ from models.weather.hourly_forecast_model import HourlyForecastModel
 
 @dataclass
 class WeatherForecastModel:
+    """Model for weather controller."""
+
     hourly: List[HourlyForecastModel]
     daily: List[DailyForecastModel]
