+++ src/models/weather/weather_forecast_data.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+"""Weather models."""
 from dataclasses import dataclass
 
 from models.weather.weather_kind import WeatherKind
@@ -6,6 +7,8 @@ from models.weather.weather_kind import WeatherKind
 
 @dataclass
 class WeatherForecastData:
+    """Weather forecast description."""
+
     kind: WeatherKind
     real_temp: float
     feels_like_temp: float
