+++ src/views/app/app_bar.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
-
+"""App views."""
 import tkinter as tk
 from dataclasses import dataclass
-from typing import List, Callable
+from typing import Callable, List
 
 from models.app.widget_kind import WidgetKind
 from views.app.app_bar_button import AppBarButtonView, AppBarButtonViewProps
@@ -11,26 +11,33 @@ from views.shared.flexible.flexible import Flexible
 
 @dataclass
 class AppBarViewProps:
+    """Props for :class:`AppBarView`."""
+
     active_tab_index: int
     tabs: List[WidgetKind]
     on_activate_tab: Callable[[int], None]
 
 
 class AppBarView(Flexible(tk.Frame)):
+    """View for application bar."""
+
     __tab_names = {
-        WidgetKind.NOTES: 'Notes',
-        WidgetKind.WEATHER: 'Weather',
-        WidgetKind.CALENDAR: 'Calendar'
+        WidgetKind.NOTES: "Notes",
+        WidgetKind.WEATHER: "Weather",
+        WidgetKind.CALENDAR: "Calendar",
     }  # TODO: интернационализация
 
     def __init__(self, parent, props: AppBarViewProps):
+        """Construct view."""
         super().__init__(parent)
         self.__props = props
         self.__render()
 
     def __render(self):
         self.grid(ipadx=4, ipady=4)
-        self.__buttons = [self.__render_button(i, t) for i, t in enumerate(self.__props.tabs)]
+        self.__buttons = [
+            self.__render_button(i, t) for i, t in enumerate(self.__props.tabs)
+        ]
 
     def __render_button(self, tab_index, tab_kind):
         btn = AppBarButtonView(
@@ -39,7 +46,7 @@ class AppBarView(Flexible(tk.Frame)):
                 text=self.__get_tab_name(tab_kind),
                 is_active=self.__props.active_tab_index == tab_index,
                 on_click=lambda: self.__props.on_activate_tab(tab_index),
-            )
+            ),
         )
         self.columnconfigure(tab_index, weight=1, pad=0)
         btn.grid(row=0, column=tab_index, ipadx=0, ipady=0, padx=0, pady=0)
@@ -48,5 +55,5 @@ class AppBarView(Flexible(tk.Frame)):
     @classmethod
     def __get_tab_name(cls, tab_kind):
         if tab_kind not in cls.__tab_names:
-            raise Exception(f'unsupported tab kind={tab_kind}')
+            raise Exception(f"unsupported tab kind={tab_kind}")
         return cls.__tab_names[tab_kind]
