+++ src/views/shared/flexible/flexible.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
+"""Flexible widget wrapper."""
 
 
 def Flexible(cls, sticky="NEWS"):
+    """Widget wrapper that makes widget flexible and applies basic grid layout."""
+
     class FlexibleWidget(cls):
         def __init__(self, *args, **kwargs):
             super().__init__(*args, **kwargs)
