+++ src/views/weather/weather.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
-
+"""Weather widget views."""
 import tkinter as tk
 
 from views.shared.flexible.flexible import Flexible
 
 
 class WeatherView(Flexible(tk.LabelFrame)):
+    """View for weather screen."""
+
     def __init__(self, parent):
-        super().__init__(parent, text='WeatherView')
+        """Construct view."""
+        super().__init__(parent, text="WeatherView")
